+++ tests/envs/test_upkie_envs.py
@@ -10,18 +10,13 @@ import unittest
 
 import gymnasium as gym
 
-import upkie.envs
-
 
 class UpkieEnvsTestCase(unittest.TestCase):
     def test_unregistered(self):
         with self.assertRaises(gym.error.NameNotFound):
             gym.make("Upkie-Servos-NotFound")
-        with self.assertRaises(gym.error.NameNotFound):
-            gym.make("Upkie-Mock-Servos")  # should not work yet
-        upkie.envs.register()
         with gym.make("Upkie-Mock-Servos") as env:
-            self.assertIsNotNone(env)  # should work now
+            self.assertIsNotNone(env)
 
 
 if __name__ == "__main__":